Problem: Filenames not standardized
Solution: Fix up filenames while adjusting tests
