+++ test/store/store_spec.js
